Cleanup dependencies + types, Remove unused Tests
Had a hell of a time trying to debug why all the tests were failing with a `Must provide document` error. Turns out, `@types/graphql` will require the latest version of `graphql`, meaning `>16`, which is where this error was coming from. To fix this, I had to restrict the peerDependencies range and set a resolutions value to make sure we're only getting a `^15.x` version of that package.
