+++ .babelrc.js
@@ -1,14 +1,12 @@
+// @ts-check
+
+/**
+ * @type {import("@babel/core").ConfigFunction}
+ */
 module.exports = {
-  plugins: [
-    require(`@babel/plugin-proposal-class-properties`),
-    require(`@babel/plugin-proposal-object-rest-spread`)
-  ],
   presets: [
     require(`@babel/preset-typescript`),
-    [
-      require(`@babel/preset-env`),
-      { targets: { node: true }, modules: false, useBuiltIns: `usage`, corejs: 3 }
-    ]
+    [require(`@babel/preset-env`), { targets: { node: true }, modules: false, useBuiltIns: `usage`, corejs: 3 }]
   ],
   env: {
     test: {
@@ -26,4 +24,4 @@ module.exports = {
       ]
     }
   }
-}
+};
