+++ api/example.ts
@@ -10,7 +10,7 @@ const typeDefs = `
 // Provide resolver functions for your schema fields
 const resolvers = {
   Query: {
-    hello: () => `Hello world!`
+    hello: (): string => `Hello world!`
   }
 };
