+++ package.json
@@ -6,8 +6,8 @@
     "GraphQL",
     "Apollo",
     "Server",
-    "Now",
-    "Javascript"
+    "Vercel",
+    "Typescript"
   ],
   "author": "Drake Costa <drake@saeris.io> (https://github.com/Saeris/)",
   "license": "MIT",
@@ -24,7 +24,7 @@
   },
   "main": "src/index.js",
   "engines": {
-    "node": ">=12.x"
+    "node": ">=12"
   },
   "scripts": {
     "ci": "yarn install --ignore-scripts --frozen-lockfile --silent --non-interactive",
@@ -42,61 +42,43 @@
     "version": "npm run build"
   },
   "dependencies": {
-    "@apollographql/graphql-playground-html": "1.6.26",
-    "@vercel/node": "^1.7.4",
-    "accept": "^3.1.3",
-    "apollo-server-core": "^2.16.1",
-    "apollo-server-env": "^2.4.5",
-    "apollo-server-types": "^0.5.1",
-    "graphql-tools": "^6.0.18"
+    "@apollographql/graphql-playground-html": "1.6.29",
+    "@vercel/node": "^1.12.1",
+    "apollo-server-core": "^2.16.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.11.1",
-    "@babel/node": "^7.10.5",
-    "@babel/plugin-proposal-class-properties": "^7.10.4",
-    "@babel/plugin-proposal-object-rest-spread": "^7.11.0",
-    "@babel/preset-env": "^7.11.0",
-    "@babel/preset-typescript": "^7.10.4",
+    "@babel/core": "^7.16.0",
+    "@babel/preset-env": "^7.16.4",
+    "@babel/preset-typescript": "^7.16.0",
     "@pika/pack": "^0.5.0",
     "@pika/plugin-build-node": "^0.9.2",
     "@pika/plugin-build-types": "^0.9.2",
     "@pika/plugin-standard-pkg": "^0.9.2",
-    "@types/accept": "^3.1.1",
+    "@saeris/eslint-config": "^2.0.2",
     "@types/graphql": "^14.5.0",
-    "@types/jest": "^27.0.2",
-    "@types/micro": "^7.3.3",
-    "@types/node": "^14.0.27",
-    "@types/supertest": "^2.0.10",
-    "@typescript-eslint/eslint-plugin": "^3.9.0",
-    "@typescript-eslint/parser": "^3.9.0",
-    "apollo-fetch": "^0.7.0",
-    "apollo-link-http": "^1.5.17",
-    "apollo-link-persisted-queries": "^0.2.2",
-    "apollo-server-plugin-response-cache": "^0.5.3",
-    "babel-core": "^6.26.3",
-    "babel-eslint": "^10.1.0",
-    "cross-env": "^7.0.2",
+    "@types/jest": "^27.0.3",
+    "@types/micro": "^7.3.6",
+    "@types/node": "^16.11.10",
+    "@types/supertest": "^2.0.11",
+    "cross-env": "^7.0.3",
     "eslint": "^7.6.0",
-    "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-jest": "^23.20.0",
-    "eslint-plugin-promise": "^4.2.1",
-    "express": "^4.17.1",
     "fiddly": "^0.12.0",
-    "graphql-subscriptions": "^1.1.0",
     "husky": "^4.2.5",
     "jest": "^27.3.1",
     "js-sha256": "^0.9.0",
-    "lint-staged": "^11.2.6",
+    "lint-staged": "^12.1.2",
     "micro": "^9.3.4",
-    "node-fetch": "^2.6.0",
-    "prettier": "2.0.5",
-    "supertest": "^4.0.2",
-    "typescript": "^3.9.7",
+    "prettier": "2.4.1",
+    "supertest": "^6.1.6",
+    "typescript": "4.4.4",
     "vercel": "^23.1.2",
     "vercel-node-server": "^2.2.1"
   },
   "peerDependencies": {
-    "graphql": "^0.12.0 || ^0.13.0 || ^14.0.0 || ^15.0.0-rc.2"
+    "graphql": "^0.12.0 || ^0.13.0 || ^14.x || >=15.x <16.0.0"
+  },
+  "resolutions": {
+    "graphql": "^15.x"
   },
   "@pika/pack": {
     "pipeline": [
