+++ src/setHeaders.ts
@@ -1,10 +1,10 @@
-import { NowResponse } from "@vercel/node"
+import type { ServerResponse } from "http";
 
-export const setHeaders = (
-  res: NowResponse,
-  headers: Record<string, any>
+export const setHeaders = <T extends ServerResponse>(
+  res: T,
+  headers: Record<string, Parameters<typeof res.setHeader>[1]>
 ): void => {
   for (const [name, value] of Object.entries(headers)) {
-    res.setHeader(name, value)
+    res.setHeader(name, value);
   }
-}
+};
