+++ src/vercelApollo.ts
@@ -1,39 +1,37 @@
-import { GraphQLOptions, HttpQueryError, runHttpQuery, convertNodeHttpToRequest } from "apollo-server-core";
-import { ValueOrPromise } from "apollo-server-types";
-import { NowRequest, NowResponse } from "@vercel/node";
+import type { GraphQLOptions, HttpQueryError } from "apollo-server-core";
+import { runHttpQuery, convertNodeHttpToRequest } from "apollo-server-core";
+import type { VercelApiHandler, VercelRequest } from "@vercel/node";
 import { setHeaders } from "./setHeaders";
 
-type NowApiHandler = (req: NowRequest, res: NowResponse) => void;
+export type VercelGraphQLOptionsFunction = (req?: VercelRequest) => GraphQLOptions | Promise<GraphQLOptions>;
 
-export interface NowGraphQLOptionsFunction {
-  (req?: NowRequest): ValueOrPromise<GraphQLOptions>;
-}
-
-export function graphqlVercel(options: GraphQLOptions | NowGraphQLOptionsFunction): NowApiHandler {
+export function graphqlVercel(options: GraphQLOptions | VercelGraphQLOptionsFunction): VercelApiHandler {
+  // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
   if (!options) throw new Error(`Apollo Server requires options.`);
 
   if (arguments.length > 1) {
     throw new Error(`Apollo Server expects exactly one argument, got ${arguments.length}`);
   }
 
-  const graphqlHandler = async (req: NowRequest, res: NowResponse) => {
+  const graphqlHandler: VercelApiHandler = async (req, res) => {
     if (req.method === `POST` && !req.body) {
-      return res.status(500).send(`POST body missing.`);
+      res.status(500).send(`POST body missing.`);
+      return;
     }
 
     try {
       const { graphqlResponse, responseInit } = await runHttpQuery([req, res], {
-        method: req.method as string,
+        method: req.method!,
         options,
-        query: req?.body || req.query,
+        query: req.body || req.query,
         request: convertNodeHttpToRequest(req)
       });
       setHeaders(res, responseInit.headers ?? {});
-      return res.status(200).send(graphqlResponse);
-    } catch (error) {
-      const { headers, statusCode, message }: HttpQueryError = error;
+      res.status(200).send(graphqlResponse);
+    } catch (error: unknown) {
+      const { headers, statusCode, message } = error as HttpQueryError;
       setHeaders(res, headers ?? {});
-      return res.status(statusCode).send(message);
+      res.status(statusCode).send(message);
     }
   };
