+++ tsconfig.eslint.json
@@ -1,0 +1,5 @@
+{
+  "extends": "./tsconfig.json",
+  "include": ["./src/**/*.*", "./api/**/*.*", "./src/**/*.d.ts"],
+  "exclude": ["node_modules"]
+}
